tests: Update test_dle for retrocompatibility
